feat: split PackageUtilities into src/utils/*
